+++ R/subscript-loc.R
@@ -25,9 +25,10 @@
 #'   vector that `i` will be matched against to construct the index. Otherwise,
 #'   not used. The default value of `NULL` will result in an error
 #'   if `i` is a character vector.
-#' @param missing Whether to throw an `"error"` when `i` is a missing
-#'   value, or `"propagate"` it (return it as is). By default, vector
+#' @param missing Whether to throw an `"error"` when `i` is a missing value,
+#'   `"propagate"` it (return it as is), or `"remove"` it. By default, vector
 #'   subscripts can contain missing values and scalar subscripts can't.
+#'   Propagated missing values can't be combined with negative indices.
 #' @param arg The argument name to be displayed in error messages when
 #'   `vec_as_location()` and `vec_as_location2()` are used to check the
 #'   type of a function input.
@@ -59,7 +60,7 @@ vec_as_location <- function(i,
                             n,
                             names = NULL,
                             ...,
-                            missing = c("propagate", "error"),
+                            missing = c("propagate", "remove", "error"),
                             arg = caller_arg(i),
                             call = caller_env()) {
   check_dots_empty0(...)
@@ -90,7 +91,7 @@ vec_as_location <- function(i,
 num_as_location <- function(i,
                             n,
                             ...,
-                            missing = c("propagate", "error"),
+                            missing = c("propagate", "remove", "error"),
                             negative = c("invert", "error", "ignore"),
                             oob = c("error", "remove", "extend"),
                             zero = c("remove", "error", "ignore"),
@@ -411,6 +412,31 @@ cnd_bullets_subscript_missing <- function(cnd, ...) {
   format_error_bullets(c(
     x = glue::glue_data(cnd, "{subscript_arg} can't contain missing values."),
     x = missing_line
+  ))
+}
+
+stop_subscript_empty <- function(i, ..., call = caller_env()) {
+  cnd_signal(new_error_subscript_type(
+    i = i,
+    body = cnd_bullets_subscript_empty,
+    ...,
+    call = call
+  ))
+}
+cnd_bullets_subscript_empty <- function(cnd, ...) {
+  cnd$subscript_arg <- append_arg("Subscript", cnd$subscript_arg)
+
+  loc <- which(cnd$i == "")
+  if (length(loc) == 1) {
+    line <- glue::glue("It has an empty string at location {loc}.")
+  } else {
+    enum <- ensure_full_stop(enumerate(loc))
+    line <- glue::glue("It has an empty string at locations {enum}")
+  }
+
+  format_error_bullets(c(
+    x = glue::glue_data(cnd, "{subscript_arg} can't contain the empty string."),
+    x = line
   ))
 }