+++ src/subscript-loc.h
@@ -8,6 +8,7 @@
 
 enum subscript_missing {
   SUBSCRIPT_MISSING_PROPAGATE = 0,
+  SUBSCRIPT_MISSING_REMOVE,
   SUBSCRIPT_MISSING_ERROR
 };
 enum num_loc_negative {