+++ tests/testthat/test-subscript-loc.R
@@ -355,6 +355,27 @@ test_that("num_as_location() with `oob = 'error'` reports negative and positive 
   })
 })
 
+test_that("num_as_location() with `missing = 'remove'` retains names (#1633)", {
+  x <- c(a = 1, b = NA, c = 2, d = NA)
+  expect_named(num_as_location(x, n = 2, missing = "remove"), c("a", "c"))
+})
+
+test_that("num_as_location() with `zero = 'remove'` retains names (#1633)", {
+  x <- c(a = 1, b = 0, c = 2, d = 0)
+  expect_named(num_as_location(x, n = 2, zero = "remove"), c("a", "c"))
+})
+
+test_that("num_as_location() with `oob = 'remove'` retains names (#1633)", {
+  x <- c(a = 1, b = 3, c = 2, d = 4)
+  expect_named(num_as_location(x, n = 2, oob = "remove"), c("a", "c"))
+})
+
+test_that("num_as_location() with `negative = 'invert'` drops names (#1633)", {
+  # The inputs don't map 1:1 to outputs
+  x <- c(a = -1, b = -3)
+  expect_named(num_as_location(x, n = 5), NULL)
+})
+
 test_that("missing values are supported in error formatters", {
   expect_snapshot({
     (expect_error(
@@ -399,6 +420,125 @@ test_that("can disallow missing values", {
       class = "vctrs_error_subscript_type"
     ))
   })
+})
+
+test_that("can alter logical missing value handling (#1595)", {
+  x <- c(a = TRUE, b = NA, c = FALSE, d = NA)
+
+  expect_identical(
+    vec_as_location(x, n = 4L, missing = "propagate"),
+    c(a = 1L, b = NA, d = NA)
+  )
+  expect_identical(
+    vec_as_location(x, n = 4L, missing = "remove"),
+    c(a = 1L)
+  )
+  expect_snapshot(error = TRUE, {
+    vec_as_location(x, n = 4L, missing = "error")
+  })
+
+  # Specifically test size 1 case, which has its own special path
+  x <- c(a = NA)
+
+  expect_identical(
+    vec_as_location(x, n = 2L, missing = "propagate"),
+    c(a = NA_integer_, a = NA_integer_)
+  )
+  expect_identical(
+    vec_as_location(x, n = 2L, missing = "remove"),
+    named(integer())
+  )
+  expect_snapshot(error = TRUE, {
+    vec_as_location(x, n = 2L, missing = "error")
+  })
+})
+
+test_that("can alter character missing value handling (#1595)", {
+  x <- c(NA, "z", NA)
+  names(x) <- c("a", "b", "c")
+  names <- c("x", "z")
+
+  expect_identical(
+    vec_as_location(x, n = 2L, names = names, missing = "propagate"),
+    set_names(c(NA, 2L, NA), names(x))
+  )
+  expect_identical(
+    vec_as_location(x, n = 2L, names = names, missing = "remove"),
+    set_names(2L, "b")
+  )
+  expect_snapshot(error = TRUE, {
+    vec_as_location(x, n = 2L, names = names, missing = "error")
+  })
+})
+
+test_that("can alter integer missing value handling (#1595)", {
+  x <- c(NA, 1L, NA, 3L)
+  names(x) <- c("a", "b", "c", "d")
+
+  expect_identical(
+    vec_as_location(x, n = 4L, missing = "propagate"),
+    x
+  )
+  expect_identical(
+    vec_as_location(x, n = 4L, missing = "remove"),
+    c(b = 1L, d = 3L)
+  )
+  expect_snapshot(error = TRUE, {
+    vec_as_location(x, n = 4L, missing = "error")
+  })
+})
+
+test_that("can alter negative integer missing value handling (#1595)", {
+  x <- c(-1L, NA, NA, -3L)
+
+  expect_snapshot(error = TRUE, {
+    num_as_location(x, n = 4L, missing = "propagate", negative = "invert")
+  })
+  expect_identical(
+    num_as_location(x, n = 4L, missing = "remove", negative = "invert"),
+    c(2L, 4L)
+  )
+  expect_snapshot(error = TRUE, {
+    num_as_location(x, n = 4L, missing = "error", negative = "invert")
+  })
+})
+
+test_that("missing value character indices never match missing value names (#1489)", {
+  x <- NA_character_
+  names <- NA_character_
+
+  expect_identical(vec_as_location(x, n = 1L, names = names, missing = "propagate"), NA_integer_)
+  expect_identical(vec_as_location(x, n = 1L, names = names, missing = "remove"), integer())
+})
+
+test_that("empty string character indices never match empty string names (#1489)", {
+  names <- c("", "y")
+
+  expect_snapshot(error = TRUE, {
+    vec_as_location("", n = 2L, names = names)
+  })
+  expect_snapshot(error = TRUE, {
+    vec_as_location(c("", "y", ""), n = 2L, names = names)
+  })
+})
+
+test_that("scalar logical `FALSE` and `NA` cases don't modify a shared object (#1633)", {
+  x <- vec_as_location(FALSE, n = 2)
+  expect_identical(x, integer())
+
+  y <- vec_as_location(c(a = FALSE), n = 2)
+  expect_identical(y, named(integer()))
+  # Still unnamed
+  expect_identical(x, integer())
+
+
+  x <- vec_as_location(NA, n = 2, missing = "remove")
+  expect_identical(x, integer())
+
+  y <- vec_as_location(c(a = FALSE), n = 2, missing = "remove")
+  expect_identical(y, named(integer()))
+  # Still unnamed
+  expect_identical(x, integer())
 })
 
 test_that("can customise subscript type errors", {